DOC: Attempt at autoreformatting docstrings to conform to Numpydoc.
I'm still working on my reformatter, but as distributed in one of the
project using numpydoc, I gave it a shot.

Most of the spacing around colons slightly change the semantic and allow
sphix/NumpyDoc to properly find the parameter names and type.

The one contentious question is whether *args, **kwargs should have */**
in front of them in docstrings (or not), and wether those should be
escaped.

Numpydoc reference does not say so but provide an example where stars
are present; and escaping seem unnecessary to render properly in sphinx.
Plus \*\* looks bad when using `?` in IPython.