Grab `func` for `finish` case only if used
The transition `func` that was being grabbed depended on whether `key`
was found in the recommendations from the last transition. If `key` was
found, one function based on `v` was called. If not, `finish` was used
in place of `v`. To make this a bit more explicit, simple, and
efficient, assign `finish` to `v` if `key` is not found. This way `v`
can be used in all cases with one call to retrieve the `func`.