Drop `ts is None` case from `_task_to_client_msgs`
This case of `ts is None` simply cannot occur as a messages cannot be
constructed to be reported. Also the handful of cases where
`_task_to_client_msgs` is called we know `ts is not None`. So there is
no point in handling this case. As such just drop the unreachable case.