Handle missing `key` case in `report_on_key` (#4755)
* Drop `ts is None` case from `_task_to_report_msg`

`ts` cannot be `None` and be properly handled here. Generally we expect
`ts is not None` and will treat it that way. Otherwise an exception will
be raised.

* Drop `ts is None` case from `_task_to_client_msgs`

This case of `ts is None` simply cannot occur as a messages cannot be
constructed to be reported. Also the handful of cases where
`_task_to_client_msgs` is called we know `ts is not None`. So there is
no point in handling this case. As such just drop the unreachable case.

* Skip gathering client messages if unneeded

If `report_msg is None`, we can skip gathering messages as there is
nothing to send. Also if there are no interested clients, there is no
need to do the work of gathering messages.

* Report `key` cancelled if missing associated task

Previously if an associated `ts` could not be found for a `key`, we
would report that `key` as cancelled. However it seems this code was
missing after a recent rewrite. This effectively readds it by
constructing a message to cancel the `key` when the task is not found.

* Test `report_on_key` with only `key`