Handle HTTPClientError in websocket connector (#4900)
Catch websocket client connection errors and re-raise them as 
CommClosedError so that they are handled correctly by the Client.

This happens when a Client tries to reconnect to a closed Cluster behind a proxy. 
The proxy closes the connection back to the client since it can't connect to the Cluster 
which leads to a HTTPClientError that gets bubbled up.