Revert "Short-circuit deps len"
This reverts commit 91aee925271e938f25a5c647a907ef8c63b74ede.

Comparing py-spy profiles for
https://github.com/gjoseph92/dask-profiling-coiled/tree/main/results#purepy-shuffle-nogc-coassign
and
https://github.com/gjoseph92/dask-profiling-coiled/tree/main/results#purepy-shuffle-nogc-coassign-short-circuit-len,
we can see that what I initially thought was a very slow `len()` call in
https://github.com/gjoseph92/dask-profiling-coiled/tree/main/results#purepy-shuffle-gc-coassign
was actually just a GC artifact. `len()` is actually just taking a tiny
amount of time in either case.

On top of that, having >5 TaskGroups as dependencies seems pretty rare,
so we probably don't need this line.