Add NannyPlugin class
This is like WorkerPlugin, but allows for code to run before the Worker
starts up.

Unfortunately this requires the Nanny to check in with the scheduler
before starting the Worker.  In principle this should be fast, but it
does delay the common case for the uncommon case.

This PR includes an Environ nanny-plugin.
If we go with this I think that we should move over PipInstall.
We might also move over UploadFile and make a new UploadDirectory