Decouple serialization, frame splitting, and compression in protocol
Currently compression and frame splitting are tightly interwoven with
traversing through messages.  This can be efficient, but results in a
complex system where it's hard to reason about when things get split or
compressed (indeed, this lead to a difficult to track down bug with
frame splitting).

This commit separates these processes into three separate stages:

1.  Serialize all objects into frames
2.  Split large frames
3.  Compress compressible frames

This results in a much more uniform application of splitting and
compressing.  However, this comes with a couple of undesired effects.

1.  We add a new header if either splitting or compressing has occurred
2.  We no longer avoid decompression when we don't want to deserialize

There is probably a clean way to achieve most/all of our goals here.
I wanted to push this up to start this conversation.