Use kwarg for Theme filename (#5190)
There was a recent change on Bokeh branch-2.4 to make the json and filename args to Theme be kwarg-only, but this is incompatible with Dask current usage. I will revert the change in Bokeh to maintain current compatibility, but I am also submitting the change here to set the stage for eventually re-introducing kwarg-only at some point in the future.

Secondarily, I have questions about what to do about testing? This was not caught by the "downstream" tests that Bokeh runs. A simple no-op test that just performs the import would have caught this since the use of Theme is at module scope. It's probably a good idea to make sure all of the bokeh-related modules cleanly import. But all of the bokeh related tests still seem to be in the dask repo, not this one. Should those tests be moved here? Or should I open a separate PR against dask to add tests?

Edit: reverted on Bokeh side at bokeh/bokeh#11484