Fix an issue where a reconnecting worker could cause an invalid transition (#5210)
If a worker reconnects with a key in memory which was already released by the scheduler, this would raise an invalid transition keyerror and would break the scheduler causing yet another deadlock.

There was also a weird issue about the `typename` function called on an instance instead of a type here
https://github.com/dask/distributed/blob/17cd4db5a86d8d78b6a4ac15a6fa41ad4e609ff6/distributed/worker.py#L922

would cause the type being submitted to be a `str` of the data object, unless it's a native object like an integer in which case it works as expected. I figured it is safer to let typename handle this situation. Very difficult to test for this particular position in the code which is why I only left another validation assert