Addressing #5159 (IPv6 dask-worker support) by propagating empty host… (#5197)
* Addressing #5159 (IPv6 dask-worker support) by propagating empty host correctly

* black / flake8 fixes

* Fixing IPv6 addresses being passed around; added tests

* Fixing rpartition differential; reverting test change

* Patching default address to match scheduler; fixing bug with fetching work from busy Worker

Prior to this commit, task results required would be fetch only from local workers if they were available. If all local workers were busy, but the work were available on another machine, this would result in an indefinite delay. This patch allows local workers to be temporarily rejected, allowing for remote workers to provide the data when all local workers are busy.

* Splitting out busy_workers into #5206

* Adding CLI test for IPv6 support

* Black fix

* Isort fix...

* Changing assertion to ValueError