Actually support uvloop in distributed
Previously distributed relied on the AnyThreadEventLoopPolicy,
which configuring uvloop would override, preventing distributed from
actually working with uvloop. We remove the need for
`AnyThreadEventLoopPolicy` and fixup uvloop configuration to fix this.

This also standardizes the `.asynchronous` and `sync` methods on classes
that support them. This is currently `Client`, `Cluster`, and
`Semaphore` for some reason. `Event`/`Lock`/... don't have these
methods, not sure why semaphore does, but I left it as is for now.