Support configuring TLS min/max version
This adds support for configuring minimum/maximum versions of TLS to
support for a connection.

Note that we also drop support for TLS versions < 1.2. These versions
are insecure, are deprecated in openSSL (and python 3.10), and shouldn't
be used. With this PR `distributed` requires TLS 1.2+. I don't
anticipate user issues from this change - TLS 1.2 was released in 2008
and should be supported everywhere at this point.