Add `distributed.client.security-loader` config
This adds a new client-side config hook for loading `Security`
configuration based on the scheduler address. This hook is only called
if no explicit `Security` object is passed to the `Client` call _and_
the client is connecting via an address (and _not_ via a `Cluster`
object). In this case the `security_loader` hook gets passed a
dictionary of information (so we can add fields later) and returns a
`Security` object to use (or `None` to fallback to the default
`Security` configuration).

This can be useful in contexts where a different security configuration
should be used to connect to different schedulers, as it lets an admin
configure a hook for selecting which certs to use when connecting to a
scheduler address without requiring the user manually do so.