Update client.scheduler_info in wait_for_workers (#5785)
Previously wait_for_workers would wait for the right number of workers,
but then the scheduler_info values were stale.  We had to wait a bit
until they updated.

But wait_for_workers gets updated info as part of it's operation.
Now we just copy this into scheduler_info as we get it.