Optionally use `mmap`ing of files with spilling
This allows the OS to handle mapping pages between file and memory. As a
result if memory is getting full and some objects are not being used,
the OS can free this memory up behind the scenes. Since the scratch
directories being used for spilling are meant to at least be local disks
(not NFS) or often solid state, IO between disk and memory can be quite
fast. So reloading spilled objects is fairly reasonable.