Pass around only bytes
This helps to reduce lots of extra unmanaged memory

This flow pretty well right now.  I'm finding that it's useful to blend
between the disk and comm buffer sizes.

The abstraction in multi_file and multi_comm are getting a little bit
worn down (it would be awkward to shift back to pandas), but maybe that's ok.