Fixup test_worker_doesnt_await_task_completion
This test was designed to verify that we don't wait for a 100s sleep
call to finish before terminating the workers.  However the time
provided, 5s, is about what it takes to restart anyway.  We now extend
this time to 20s, which is safely below 100s, and also safely above the
worker start time.

In the future we should figure out why it's taking us 5s to start up.
This is drifting upwards