Use broader range in test_nanny_worker_port_range (#5980)
This test is designed to verify that we can restrict workers and nannies
to a range of ports.  It creates three workers and gives three ports.

However, if one of those ports are already taken then this test will
time out.  Unfortunately our cleanup is not perfect, and sometimes ports
are taken, leading to flaking tests.

This commit changes the behavior of the test by giving a much broader
range of port values, but still narrow enough that it is unlikely to
pass by chance.