Relax variable test_race
The intent of this test is mostly to make sure that lots of workers can
access the same variable in parallel without breaking things.

Part of this test also ensures that none of them got too far behind the
others. This is very hard to actually guarantee though.  There was
some slippage built into the test, but a particularly long wait in one
of the workers can set the entire thing back arbitrarily far.  We
remove the extra part of this test.