Convert test_restart_sync_no_center
This test is ANCIENT.  "Center" was the original term for "Scheduler"
back when only I was writing on this thing.

Anyway, this test was behaving oddly.  I'm shifting it to a gen_cluster
test and I suspect that that will either help things to work, or at
least will give us a lot more visibility about why it isn't working.