Maybe fix test_worker_waits_for_scheduler
Previously this was timing out in some test runs.
One *possible* issue is that we started a coroutine to ask the
scheduler to release us, but the scheduler doesn't exist.  This
removes that request.

I'm not confident that this will fix the issue.  I give around a 30%
chance.  However, I am fairly confident that this won't have negative
consequences.