Move shuffle worker assignment logic to the scheduler
Previously we assigned a partition to a worker based on a simple
formula.  This was good, but also error prone in a few ways:

1.  We would send data around needlessly if the task wasn't desired
2.  Conflicting shuffles could assign tasks to conflicting workers

Now we as the scheduler to manage this process, and all of the workers
check in with it to get assignments.  Currently the actual logic is the
same, but now things get to live in a single location where we can make
smarter decisions to avoid conflicts.

This removes the previous shuffle setup steps on the worker, simplifying
the code a bit (subjectively anyway).  It makes the compute steps a bit
worse because now we have a small pandas join in the middle of things.
It should already avoid excess labor.  It does not yet avoid restriction
conflicts.