Drop `cast`ing which is effectively a no-op
In both of these cases `mv` is simply being cast to a `format` (and
`shape`) that it already is. Though this was an accidental overwrite by
a previous bug fix PR. The code previously tried to maintain `mv`'s
`format` and `shape` before performing a copy. While that behavior could
be restored, the current behavior has been unchanged for over a year.
This indicates the `cast`ing was likely unnecessary to begin with. So
just drop it. If we find we need it after all, it is easy to add
correctly, which require different code than what is here anyways.