Use async with Server in test_core.py
Previously we would create and rarely clean up servers in tests.
Now we use the async context manager whenever it is easy to do.

This was inspired by the following CI test failure.

https://github.com/dask/distributed/runs/5962479318?check_suite_focus=true

This doesn't necessarily solve the problem, but it was easy to do and I doubt that it hurts.