Update `locket._LockFile` implementation to avoid leaving open files
The `LockFile` implementation opens a file descriptor to check locks. In the non-blocking case, if the lock cannot be acquired, a `LockError` is thrown and an open file descriptor is left dangling. In normal operation, this file is closed on lock release, but since the lock has already been acquired by another instance of the `LockFile`, release never occurs.