Remove legacy state mappings
These are a historical artifact of how we used to manage state in the
scheduler.  They were mostly around because lots of old tests depended
on this state.  I've gone through and manually updated all of the tests.

This PR is also a good example of what PRs to distributed used to look
like before we switched away from precise state-driven tests.
Thank goodness that 90% of our tests no longer dive into internal state.