Pass on in-flight transfers if they are already released
A task that is released while still in-flight will stick around in the
in_flight_workers state.  This can cause a KeyError when gather_dep goes
to handle this now-missing-task.

The correct thing to do in this case is to just ignore the key,
which is what this commit does.

We could also keep in_flight_workers up-to-date on a release transition.
I'm not sure how valuable this would be apart from passing.
I'm open to either approach.  This was the easiest.

Fixes https://github.com/dask/distributed/issues/6194