Explicitly require python-snappy 0.5.3+
Ideally we would test `__version__` however not in a release atm (though
it was added recently). In any even it can't reliably be checked for
current versions. Instead just test some fast code that reliably fails
for versions prior to `0.5.3`. If the code fails, simply raise an
`ImportError` noting this requirement. As this is in a `suppress` block,
the `ImportError` will be caught. Though the net effect will be
disabling python-snappy use if it is too old.