Use `memoryview` of `payload`
To allow more efficient accessing of the `payload` (like when selecting
portions in `byte_sample`), take a `memoryview` of the data. Ensure that
is 1-D contiguous `uint8` data. This makes it very similar to `bytes`,
which will work well in `byte_sample` and compressors that handle only a
narrow form of the Python Buffer Protocol.

This allows us to drop various `ensure_bytes` calls in compression that
would otherwise copy the data. Should reduce memory usage when
serializing as part of transmission or spilling.