Add scheduler-sni option for dask workers
This adds support for dask workers to specify the SNI host
separately from the TLS endpoint.  In some environments (such
as a Kubernetes cluster running dask-gateway), the schedulers
may be behind a TLS proxy, while the workers are external to the
cluster.  This allows workers to connect to the schedulers
without the need for wildcard DNS.

Signed-off-by: Burt Holzman <burt@fnal.gov>