Coerce `frame` to `bytes` directly
Instead of relying on `ensure_bytes` in these code paths (where we know
types like `str` are not possible), simply check if each `frame` is
`bytes` and if not construct a `bytes` object from the `frame`. Since
`bytes` constructor can take anything that follows the Python Buffer
Protocol, it can coerce `bytearray`s, `memoryview`s, etc. without issue.
This simplifies some of the handling here.

Should note that we really do need to coerce each `frame` to `bytes` as
Tornado only supports `bytes` objects in these code paths. Though it may
be possible to relax that constraint, that would need to happen in
Tornado itself. So continue with the status quo for now of copying each
`frame` to `bytes` when needed.