Track the size of all frames sent while sending
Instead of having a second `for`-loop go back over the frames and
compute how much data was sent, just compute that as part of the frame
sending process. This is also trivial to do after each `frame` has been
coerced to `bytes` as the number of bytes sent is simply the `frame`
length. Should remove some unnecessary overhead.