Wait on worker execution until plugins have started
Previously a worker was technically able to accept tasks before it was
in a started state.  This uses the `_started` Event on the Server class
and avoids execution until that is set.

This could be improved.  Ideally we wouldn't consider really accepting
tasks at all until we were in a started state.  This is a good first
step though.