Don't queue tasks on workers
With this change, workers are not sent more tasks than they have threads. This eliminates root task overproduction and removes the need for work stealing.

This is optimized for a minimal diff. There is _lots_ we might change/rip out if we went forward with this.

Performance will likely be poor on some workloads without speculative task assignment and root task co-assignment.