Raise for empty `memoryview` w/multiple dimensions
It is not possible to correctly deserialize an empty `memoryview` with
multiple dimensions using builtin Python functionality alone. So raise
when trying to serialize such an object to provide this information the
user as close to the error producing object as possible (and not when it
is deserialized somewhere else). Should help the user find this and
either flatten the `memoryview` themselves, switching to using NumPy
where this is handled correctly, or do something else.